Notebooks to optimize sailor health
Based on the scoring mechanism, the team's goal was to keep all the sailors alive, with health score between 1 and 100, for days 1 and 2. For day 3, the goal was to use all the medication in order to maximize health. The notebooks achieve this using a "nearly greedy" algorithm:  Any sailor with health needs over 25 received Remdesivir (+30 health), with needs over 20 received Dexamethasone (+25), over 15 received Plasma and Chloroquine (+10 each), over 10 received Casirivimab (+15). Those with health needs between 1 and 10 received Plasma on Day 1 and Chloroquine on Day 2. On Day 3, the algorithm distributed the strongest available medicine to the sickest sailor, then recalculated expected health and repeated.